refactors transform and adds serializable transform
